Some 64-bit fixes. Found some headers that were being copied twice.

+++ Source/DOM classes/Core DOM/NamedNodeMap.m
@@ -87,7 +87,7 @@
 
 -(unsigned long)length
 {
-	int count = [self.internalDictionary count];
+	NSUInteger count = [self.internalDictionary count];
 	
 	for( NSDictionary* namespaceDict in [self.internalDictionaryOfNamespaces allValues] )
 	{
